requirements.txt
astroid==2.0.4
certifi==2018.11.29
Django==2.1.7
django-cors-headers==2.4.0
django-filter==2.1.0
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
drf-extensions==0.4.0
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.1.1
pytz==2018.9
wrapt==1.10.11
mysqlclient==1.4.2